# docs/openapi.yaml
openapi: 3.1.0
info:
  title: hodei-pipelines-server
  description: Hodei Pipelines Server - Monolithic Modular Architecture
  contact:
    name: Hodei Team
  license:
    name: MIT
    identifier: MIT
  version: 0.16.0
paths:
...
  /api/v1/logs/statistics:
    get:
      tags:
      - observability
      summary: GET /api/v1/logs/statistics - Get log statistics
      operationId: logs_statistics_handler
      parameters:
      - name: tenant_id
        in: query
        description: Filter by tenant ID
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Log statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogStatistics'
        '500':
          description: Internal server error
...
  /api/v1/pipelines/{id}:
    get:
      tags:
      - pipelines
      operationId: get_pipeline_handler
      parameters:
      - name: id
        in: path
        description: Pipeline ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Pipeline details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDto'
        '404':
          description: Pipeline not found
        '500':
          description: Internal server error
    put:
      tags:
      - pipelines
      operationId: update_pipeline_handler
      parameters:
      - name: id
        in: path
        description: Pipeline ID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePipelineRequestDto'
        required: true
      responses:
        '200':
          description: Pipeline updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDto'
        '404':
          description: Pipeline not found
        '500':
          description: Internal server error
    delete:
      tags:
      - pipelines
      operationId: delete_pipeline_handler
      parameters:
      - name: id
        in: path
        description: Pipeline ID
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Pipeline deleted successfully
        '500':
          description: Internal server error
...
components:
  schemas:
...
    CreatePipelineStepRequestDto:
      type: object
      required:
      - name
      - job_spec
      - dependencies
      properties:
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        id:
          type:
          - string
          - 'null'
          format: uuid
        job_spec:
          $ref: '#/components/schemas/JobSpecDto'
        name:
          type: string
...
    JobSpecDto:
      type: object
      required:
      - name
      - image
      - command
      - resources
      - timeout_ms
      - retries
      - env
      - secret_refs
      properties:
        command:
          type: array
          items:
            type: string
        env:
          type: object
          additionalProperties:
            type: string
          propertyNames:
            type: string
        image:
          type: string
        name:
          type: string
        resources:
          $ref: '#/components/schemas/ResourceQuotaDto'
        retries:
          type: integer
          format: int32
          minimum: 0
        secret_refs:
          type: array
          items:
            type: string
        timeout_ms:
          type: integer
          format: int64
          minimum: 0
...
    LogStatistics:
      type: object
      description: Log statistics structure
      required:
      - total_logs
      - by_log_level
      - by_time_period
      - top_search_terms
      - error_rate
      - timestamp
      properties:
        by_log_level:
          type: object
          description: Logs by log level
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
          propertyNames:
            type: string
        by_time_period:
          type: object
          description: Logs by time period
          additionalProperties:
            type: integer
            format: int64
            minimum: 0
          propertyNames:
            type: string
        error_rate:
          type: number
          format: double
          description: Error rate percentage
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the statistics
        top_search_terms:
          type: array
          items:
            type: string
          description: Top search terms
        total_logs:
          type: integer
          format: int64
          description: Total number of logs
          minimum: 0
...
    PipelineDto:
      type: object
      required:
      - id
      - name
      - status
      - steps
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
          format: date-time
        description:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/PipelineStepDto'
        updated_at:
          type: string
          format: date-time
    PipelineStepDto:
      type: object
      required:
      - id
      - name
      - job_spec
      - dependencies
      properties:
        dependencies:
          type: array
          items:
            type: string
            format: uuid
        id:
          type: string
          format: uuid
        job_spec:
          $ref: '#/components/schemas/JobSpecDto'
        name:
          type: string
...
    ResourceQuotaDto:
      type: object
      required:
      - cpu_m
      - memory_mb
      properties:
        cpu_m:
          type: integer
          format: int64
          minimum: 0
        gpu:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        memory_mb:
          type: integer
          format: int64
          minimum: 0
...
    UpdatePipelineRequestDto:
      type: object
      properties:
        description:
          type:
          - string
          - 'null'
        name:
          type:
          - string
          - 'null'
        steps:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/CreatePipelineStepRequestDto'
        variables:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
...